misc changes

+++ database/migrations/2016_12_16_125614_create_options_house_transaction_table.php
@@ -23,7 +23,7 @@ class CreateOptionsHouseTransactionTable extends Migration
             $table->time('close_time');
             $table->string('trade_type');
             $table->string('description');
-            $table->integer('strike_price');
+            $table->float('strike_price');
             $table->string('option_type');
             $table->string('option_side');
             $table->integer('option_quantity')->nullable();
